requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Database
motor==3.3.2
pymongo==4.6.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# File Upload & Processing
python-multipart==0.0.6

# Optional: Additional packages for marketplace features
# cloudinary==1.36.0
# stripe==7.8.0
# twilio==8.10.0
# redis==5.0.1
# celery==5.3.4
# Pillow==10.1.0